test: iconify icons
