+++ packages/starlight/__e2e__/fixtures/basics/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@astrojs/starlight": "workspace:*",
+    "@iconify-json/mdi": "^1.2.3",
     "astro": "^5.1.5"
   }
 }
