+++ packages/starlight/__e2e__/fixtures/basics/src/content/docs/iconify.mdx
@@ -1,0 +1,31 @@
+---
+title: Iconify
+---
+
+import TestId from '../../components/TestId.astro';
+import { Aside, Card, Icon } from '@astrojs/starlight/components';
+
+<TestId id="icon-component">
+	<Icon name="mdi:checkbox-blank-circle" />
+</TestId>
+
+<TestId id="card-component">
+	<Card icon="mdi:checkbox-blank-circle" title="Example">
+		A card with an icon from the user-installed Iconify `mdi` set.
+	</Card>
+</TestId>
+
+<TestId id="aside-component">
+	<Aside icon="mdi:checkbox-blank-circle">
+		Aside with an icon from the user-installed Iconify `mdi` set.
+	</Aside>
+</TestId>
+
+
+<TestId id="aside-directive">
+
+:::note{icon="mdi:checkbox-blank-circle"}
+Aside with an icon from the user-installed Iconify `mdi` set.
+:::
+
+</TestId>
